revert unknown params
